Repair Inventory_DialogWindow complete with class Inventory(now is not one object) with class Item (work with std::shared_ptr<Item>, old version Item*) repair functions

+++ Silently/src/RPGspace/Inventory/inventory.cpp
@@ -17,7 +17,7 @@ QVector<std::shared_ptr<Item>> Inventory::getItems() const{
     return items_;
 }
 
-size_t Inventory::getSize() const{
+int Inventory::getSize() const{
     if(items_.isEmpty()){
         return 0;
     }
